channel deletion: move impl details to channel.go
