+++ go/src/socialapi/models/channel_bongo.go
@@ -5,8 +5,6 @@ import (
 	"time"
 
 	"github.com/koding/bongo"
-
-	"socialapi/request"
 )
 
 const ChannelTableName = "api.channel"
@@ -93,89 +91,4 @@ func (c *Channel) FetchByIds(ids []int64) ([]Channel, error) {
 
 func (c *Channel) CountWithQuery(q *bongo.Query) (int, error) {
 	return bongo.B.CountWithQuery(c, q)
-}
-
-func getMessageBatch(channelId int64, c int) ([]ChannelMessage, error) {
-	messageIds, err := NewChannelMessageList().
-		FetchMessageIdsByChannelId(channelId, &request.Query{
-		Skip:  c * 100,
-		Limit: 100,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return NewChannelMessage().FetchByIds(messageIds)
-}
-
-func isMessageCrossIndexed(messageId int64) (error, bool) {
-	count, err := NewChannelMessageList().CountWithQuery(&bongo.Query{
-		Selector: map[string]interface{}{
-			"message_id": messageId,
-		},
-	})
-	if err != nil {
-		return err, false
-	}
-	return nil, count > 1
-}
-
-func (c *Channel) deleteChannelMessages() error {
-	if c.Id == 0 {
-		return ErrIdIsNotSet
-	}
-	for i := 0; ; i++ {
-		messages, err := getMessageBatch(c.Id, i)
-		if err != nil {
-			return err
-		}
-		for _, message := range messages {
-			err, isCrossIndexed := isMessageCrossIndexed(message.Id)
-			if err != nil {
-				return err
-			}
-
-			if isCrossIndexed {
-				continue
-			}
-
-			if err = message.Delete(); err != nil {
-				return err
-			}
-		}
-		if len(messages) < 100 {
-			return nil
-		}
-	}
-}
-
-func getListingBatch(channelId int64, c int) ([]ChannelMessageList, error) {
-	var listings []ChannelMessageList
-	q := &bongo.Query{
-		Selector: map[string]interface{}{"channel_id": channelId},
-		Pagination: bongo.Pagination{
-			Skip:  100 * c,
-			Limit: 100,
-		}}
-	if err := NewChannelMessageList().Some(&listings, q); err != nil {
-		return nil, err
-	}
-	return listings, nil
-}
-
-func (c *Channel) deleteChannelLists() error {
-	for i := 0; ; i++ {
-		listings, err := getListingBatch(c.Id, i)
-		if err != nil {
-			return err
-		}
-
-		for _, listing := range listings {
-			if err := listing.Delete(); err != nil {
-				return err
-			}
-		}
-		if len(listings) < 100 {
-			return nil
-		}
-	}
 }
